[Feat] : Chat, Keyword 삭제 api 연동

+++ app/src/main/java/com/example/onenthapp/data/chat/ChatNameResponse.kt
@@ -9,5 +9,5 @@ data class ChatNameResponse(
 
 data class ChatNameResult(
     val chatRoomName: String,
-    val chatRoomId: Long
+    val chatRoomId: Int
 )
